Use mock-cart-api in cart-actions tests
