+++ packages/shopping-cart/test/cart-actions.tsx
@@ -8,11 +8,17 @@ import {
 import React, { useEffect, useRef } from 'react';
 import { getEmptyResponseCart } from '../src/empty-carts';
 import { useShoppingCart, withShoppingCart, ShoppingCartProvider } from '../src/index';
+import {
+	getCart,
+	setCart,
+	planOne,
+	planTwo,
+	renewalOne,
+	renewalTwo,
+	mainCartKey,
+} from './utils/mock-cart-api';
 import type {
 	RequestCartProduct,
-	ResponseCartProduct,
-	RequestCart,
-	ResponseCart,
 	MinimalRequestCartProduct,
 	GetCart,
 	SetCart,
@@ -24,122 +30,7 @@ import type {
 import '@automattic/calypso-polyfills';
 import '@testing-library/jest-dom/extend-expect';
 
-const planOne: ResponseCartProduct = {
-	time_added_to_cart: Date.now(),
-	current_quantity: 1,
-	product_name: 'WordPress.com Personal',
-	product_slug: 'personal-bundle',
-	currency: 'BRL',
-	extra: {
-		context: 'signup',
-	},
-	meta: '',
-	product_id: 1009,
-	volume: 1,
-	quantity: null,
-	item_original_cost_integer: 14400,
-	item_original_cost_display: 'R$144',
-	item_subtotal_integer: 14400,
-	item_subtotal_display: 'R$144',
-	product_cost_integer: 14400,
-	product_cost_display: 'R$144',
-	item_subtotal_monthly_cost_display: 'R$144',
-	item_subtotal_monthly_cost_integer: 14400,
-	item_original_subtotal_integer: 14400,
-	item_original_subtotal_display: 'R$144',
-	is_domain_registration: false,
-	is_bundled: false,
-	is_sale_coupon_applied: false,
-	months_per_bill_period: null,
-	uuid: 'product001',
-	cost: 144,
-	price: 144,
-	product_type: 'test',
-	included_domain_purchase_amount: 0,
-};
-
-const planTwo: ResponseCartProduct = {
-	time_added_to_cart: Date.now(),
-	current_quantity: 1,
-	product_name: 'WordPress.com Business',
-	product_slug: 'business-bundle',
-	currency: 'BRL',
-	extra: {
-		context: 'signup',
-	},
-	meta: '',
-	product_id: 1010,
-	volume: 1,
-	quantity: null,
-	item_original_cost_integer: 14400,
-	item_original_cost_display: 'R$144',
-	item_subtotal_integer: 14400,
-	item_subtotal_display: 'R$144',
-	product_cost_integer: 14400,
-	product_cost_display: 'R$144',
-	item_subtotal_monthly_cost_display: 'R$144',
-	item_subtotal_monthly_cost_integer: 14400,
-	item_original_subtotal_integer: 14400,
-	item_original_subtotal_display: 'R$144',
-	is_domain_registration: false,
-	is_bundled: false,
-	is_sale_coupon_applied: false,
-	months_per_bill_period: null,
-	uuid: 'product002',
-	cost: 144,
-	price: 144,
-	product_type: 'test',
-	included_domain_purchase_amount: 0,
-};
-
-const renewalOne: ResponseCartProduct = {
-	...planOne,
-	extra: { purchaseType: 'renewal' },
-};
-
-const renewalTwo: ResponseCartProduct = {
-	...planTwo,
-	extra: { purchaseType: 'renewal' },
-};
-
-const mainCartKey = '1';
-
 const emptyResponseCart = getEmptyResponseCart();
-
-async function getCart( cartKey: string ) {
-	if ( cartKey === mainCartKey ) {
-		return emptyResponseCart;
-	}
-	throw new Error( 'Unknown cart key' );
-}
-
-function createProduct( productProps: RequestCartProduct ): ResponseCartProduct {
-	switch ( productProps.product_id ) {
-		case 1009:
-			return planOne;
-		case 1010:
-			return planTwo;
-	}
-	throw new Error( 'Unknown product' );
-}
-
-async function setCart( cartKey: string, newCart: RequestCart ): Promise< ResponseCart > {
-	if ( [ 'no-site', 'no-user', mainCartKey ].includes( cartKey ) ) {
-		// Mock the shopping-cart endpoint response here
-		return {
-			...emptyResponseCart,
-			cart_key: cartKey,
-			products: newCart.products.map( createProduct ),
-			coupon: newCart.coupon,
-			is_coupon_applied: !! newCart.coupon,
-			tax: {
-				display_taxes: !! newCart.tax?.location.postal_code,
-				location: newCart.tax?.location,
-			},
-		};
-	}
-	throw new Error( 'Unknown cart key' );
-}
 
 function ProductList( {
 	initialProducts,
